+++ Sources/Graphics/Graphics.h
@@ -22,8 +22,8 @@ namespace Graphics {
     LIB_API void DrawPoint(const Point& p, float pointSize, const Color4f& color);
     LIB_API void DrawPoints(const std::vector<Point>& v, float pointSize, const Color4f& color);
     LIB_API void DrawLines(const std::vector<Point>& v, const Color4f& color);
-    LIB_API float win2worldX(float x);
-    LIB_API float win2worldY(float y);
-    LIB_API float world2winX(float x);
-    LIB_API float world2winY(float y);
+    LIB_API float win2worldX(int x);
+    LIB_API float win2worldY(int y);
+    LIB_API int world2winX(float x);
+    LIB_API int world2winY(float y);
 }
